+++ lua/miriani/scripts/games.lua
@@ -6,36 +6,48 @@ ImportXML([=[
   <trigger
    enabled="y"
    group="games"
-   match="^You pick up the silver laser gun attached to the game console and begin a game of Chicken Chase!$"
+   match="^.+? picks? up the silver laser gun attached to the game console and begins? a game of Chicken Chase!$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_start", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_start", "games")</send>
   </trigger>
 
 
   <trigger
    enabled="y"
    group="games"
-   match="^A[n]? .+? chicken .+? screen\. You take careful aim with the laser gun and fire\.$"
+   match="^A[n]? .+? chicken .+? screen\. .+? (?:aims the gun controller at it and|take careful aim with the laser gun and) fires?\.$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_shoot", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_shoot", "games")</send>
   </trigger>
 
   <trigger
    enabled="y"
    group="games"
-   match="^A Chicken Chase game announces, &quot;Points\: \d+\.  Shots left: \d+\.&quot;$"
+   match="^A Chicken Chase game announces, &quot;.+?&quot;$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_announce", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_announce", "games")</send>
   </trigger>
+
+  <trigger
+   enabled="y"
+   group="games"
+   match="^.+? laser shot missed entirely!$"
+   regexp="y"
+   send_to="12"
+   sequence="100"
+  >
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_miss", "games")</send>
+  </trigger>
+
 
   <trigger
    enabled="y"
@@ -45,51 +57,51 @@ ImportXML([=[
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_miss", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_miss", "games")</send>
   </trigger>
 
   <trigger
    enabled="y"
    group="games"
-   match="^Your laser barely hits the chicken's tail feathers\. That's worth \d+ points?\.$"
+   match="^.+? barely hits the chicken's tail feathers\. That's worth \d+ points?\.$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_hit", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_hit", "games")</send>
   </trigger>
 
   <trigger
    enabled="y"
    group="games"
-   match="^Your laser shot clips one foot as the chicken flaps across the screen\. \d+ points?\.$"
+   match="^.+? laser shot clips one foot as the chicken flaps across the screen\. \d+ points?\.$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_hit", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_hit", "games")</send>
   </trigger>
 
   <trigger
    enabled="y"
    group="games"
-   match="^The red laser shot hits the center of one wing!  \d+ points?!$"
+   match="^.+? red laser shot hits the center of one wing.  ?\d+ points?!$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_hit", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_hit", "games")</send>
   </trigger>
 
   <trigger
    enabled="y"
    group="games"
-   match="^You nail a solid body shot for \d+ points\. The chicken lets out a loud squawk\.$"
+   match="^.+? nails? a solid body shot.*?\. The chicken lets out a loud squawk\..*?$"
    regexp="y"
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_squawk", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_squawk", "games")</send>
   </trigger>
 
   <trigger
@@ -100,7 +112,18 @@ ImportXML([=[
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_explode", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_explode", "games")</send>
+  </trigger>
+
+  <trigger
+   enabled="y"
+   group="games"
+   match="^The chicken explodes in a burst of feathers as .+? laser shot nails it in the head!$"
+   regexp="y"
+   send_to="12"
+   sequence="100"
+  >
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_explode", "games")</send>
   </trigger>
 
   <trigger
@@ -111,7 +134,59 @@ ImportXML([=[
    send_to="12"
    sequence="100"
   >
-  <send>mplay("misc/chicken_chase_game_over", "games", 1)</send>
+  <send>mplay("misc/games/chicken chase/chicken_chase_game_over", "games")</send>
+  </trigger>
+
+ <trigger
+   enabled="y"
+   group="games"
+   match="^.+? winds? up and sends? a ball down the lane\.(?:  Good luck\.)?$"
+   regexp="y"
+   send_to="12"
+  >
+   <send>mplay("misc/games/skeeball/skeballRoll", "games")</send>
+  </trigger>
+
+  <trigger
+   enabled="y"
+   group="games"
+   match="^A .+? skeeball machine announces, &quot;(?:.+? has \d+ points? with \d+ balls? remaining|Points: \d+\.  Balls left: \d+)\.&quot;$"
+   regexp="y"
+   send_to="12"
+  >
+   <send>mplay("misc/games/skeeball/skeballScore", "games")</send>
+  </trigger>
+
+  <trigger
+   enabled="y"
+   group="games"
+   match="^Game over\. .+? score (?:is|was) \d+(?: points?\.)?$"
+   regexp="y"
+   send_to="12"
+  >
+   <send>mplay("misc/games/skeeball/skeballEnd", "games")</send>
+  </trigger>
+
+<trigger
+   enabled="y"
+   group="games"
+   match="After a few seconds, the sound of ravens cawing plays from a hidden speaker on a spiderweb-covered pyramidal machine as if in mockery\.$"
+   regexp="y"
+   send_to="12"
+   sequence="100"
+  >
+  <send>mplay("misc/games/prize machines/cawingRavens", "games")</send>
+  </trigger>
+
+  <trigger
+   enabled="y"
+   group="games"
+   match="After a few seconds, a faint thud can be heard, and .+ falls into the retrieval slot at the bottom of a spiderweb-covered pyramidal machine\.(?: [A-Z][^.]{1,80}\.)?$"
+   regexp="y"
+   send_to="12"
+   sequence="100"
+  >
+  <send>mplay("misc/games/prize machines/slotPayout", "games")</send>
   </trigger>