+++ lua/miriani/scripts/shenanigans.lua
@@ -236,27 +236,6 @@ ImportXML([=[
   <send>mplay("activity/archaeology/practice")</send>
   </trigger>
 
-<trigger
-   enabled="y"
-   group="shenanigans"
-   match="After a few seconds, the sound of ravens cawing plays from a hidden speaker on a spiderweb-covered pyramidal machine as if in mockery\.$"
-   regexp="y"
-   send_to="12"
-   sequence="100"
-  >
-  <send>mplay("misc/cawingRavens")</send>
-  </trigger>
-
-  <trigger
-   enabled="y"
-   group="shenanigans"
-   match="After a few seconds, a faint thud can be heard, and .+ falls into the retrieval slot at the bottom of a spiderweb-covered pyramidal machine\.(?: [A-Z][^.]{1,80}\.)?$"
-   regexp="y"
-   send_to="12"
-   sequence="100"
-  >
-  <send>mplay("misc/slotPayout")</send>
-  </trigger>
 
   <!-- Air Freshener Sounds -->