+++ docs/changelog.md
@@ -4,6 +4,9 @@
 ### Fixed
 - Fix ambiance not immediately following power or config changes.
 
+### New
+- Added ambiance for ferries.
+
 ## Version 4.1.16
 ### New
 - Added a zap for witnessing someone pick up a sanitation drone. Now it literally hurts to watch.